+++ src/slogo/model/movementcommand/SetTowards.java
@@ -26,7 +26,7 @@ public class SetTowards implements Command {
     }
 
     @Override
-    public int execute(Turtle turtle) {
+    public double execute(Turtle turtle) {
         double xdiff = x - turtle.getXCoordinate();
         double ydiff = y - turtle.getYCoordinate();
         int degree;
